refactor(core): longPolling
